SETI-995: Install custom version of python-jenkins
The code for retries on server errors is not yet merged upstream:
https://review.openstack.org/522804

I also had to add explicitly the requirements.txt to deps in tox.ini,
otherwise tox would install the upstream python-jenkins. This is
caused by tox running following command:
  pip -e $PWD

which in turn runs:
  python setup.py egg_info

which writes requirements file with `python_jenkins`. It looks like
setuptools/pbr does not know how to work with git dependencies.